requirements.txt
tensorflow==2.17.0
faiss-cpu==1.8.0.post1